+++ src/common/camera/index.ts
