+++ src/common/index.ts
@@ -1,0 +1,5 @@
+import camera from './camera';
+
+export default {
+  camera,
+};
