+++ src/renderer/AbstractImageRenderer/index.ts
