+++ src/renderer/EdgeImageRenderer/glsl/fs.glsl
