+++ src/renderer/EdgeImageRenderer/glsl/vs.glsl
