+++ src/renderer/EdgeImageRenderer/index.ts
