+++ src/renderer/MagnifierImageRenderer/glsl/fs.glsl
