+++ src/renderer/MagnifierImageRenderer/glsl/vs.glsl
