+++ src/renderer/MagnifierImageRenderer/index.ts
