+++ src/renderer/SharpenImageRenderer/index.ts
