+++ src/renderer/SmoothImageRenderer/glsl/fs.glsl
