+++ src/renderer/SmoothImageRenderer/glsl/vs.glsl
