+++ src/renderer/SmoothImageRenderer/index.ts
