--- src/rendering/core/index.ts
@@ -1,5 +1,0 @@
-import Shader from './Shader';
-
-export default {
-  Shader,
-};
