--- src/rendering/index.ts
@@ -1,11 +1,0 @@
-import core from './core';
-import camera from './camera';
-import renderer from './renderer';
-import window from './window';
-
-export default {
-  core,
-  camera,
-  renderer,
-  window,
-};
