+++ src/window/EdgeImageWindow/index.ts
