+++ src/window/MagnifierImageWindow/index.ts
