+++ src/window/SharpenImageWindow/index.ts
