+++ src/window/SmoothImageWindow/index.ts
