+++ src/window/index.ts
